MOD: code connect
main merge 후 기능 연결 작업 완료
