+++ app/src/main/AndroidManifest.xml
@@ -16,13 +16,16 @@
         tools:targetApi="31">
 
         <activity
-            android:name=".SplashActivity"
+            android:name=".BookclubBookReview"
             android:exported="false" />
         <activity
-            android:name=".bookclub_book.BookClubJoinActivity"
+            android:name=".BookclubJoin"
             android:exported="false" />
         <activity
-            android:name=".setting.SettingActivity"
+            android:name=".MyPageWriteReview"
+            android:exported="false" />
+        <activity
+            android:name=".Setting"
             android:exported="false" />
         <activity          
             android:name=".Preference4"
@@ -59,9 +62,6 @@
             android:exported="false" />
         <activity
             android:name=".TermsofUse"
-            android:exported="true" />
-        <activity
-            android:name=".bookclub_book.BookclubBookActivity"
             android:exported="true" />
         <activity
             android:name=".MainActivity"
