--- app/src/main/java/com/example/fe/MypageFragment.kt
@@ -1,23 +1,0 @@
-package com.example.fe
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.example.fe.databinding.FragmentMypageBinding
-
-class MypageFragment : Fragment() {
-
-    private lateinit var binding: FragmentMypageBinding
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        binding = FragmentMypageBinding.inflate(inflater, container, false)
-
-        return binding.root
-    }
-}
