+++ app/src/main/java/com/example/fe/SplashActivity.kt
@@ -6,21 +6,19 @@ import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
 import androidx.appcompat.app.AppCompatActivity
-import com.example.fe.bookclub_book.BookclubBookActivity
 
 @SuppressLint("CustomSplashScreen")
 class SplashActivity:AppCompatActivity() {
-
-    private val SPLASH_TIME_OUT: Long = 2000 // 스플래시 화면 지속 시간 (3초)
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_splash)
-
-        Handler(Looper.getMainLooper()).postDelayed({
-            // MainActivity로 전환
-            startActivity(Intent(this, BookclubBookActivity::class.java))
-            finish()
-        }, SPLASH_TIME_OUT)
+//
+//    private val SPLASH_TIME_OUT: Long = 2000 // 스플래시 화면 지속 시간 (3초)
+//
+//    override fun onCreate(savedInstanceState: Bundle?) {
+//        super.onCreate(savedInstanceState)
+//        setContentView(R.layout.activity_splash)
+//
+//        Handler(Looper.getMainLooper()).postDelayed({
+//            // MainActivity로 전환
+//            startActivity(Intent(this, BookclubBookActivity::class.java))
+//            finish()
+//        }, SPLASH_TIME_OUT)
     }
-}
