+++ app/src/main/java/com/example/fe/bookclub_book/BookclubBookHomeFragment.kt
@@ -2,21 +2,19 @@ package com.example.fe.bookclub_book
 
 import android.content.Intent
 import android.os.Bundle
-import android.telecom.Call
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.view.WindowInsetsAnimation
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.commit
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.example.fe.BookclubJoin
+import com.example.fe.BookclubBookReview
 import com.example.fe.bookclub_book.dataclass.BookclubByMonth
 import com.example.fe.bookclub_book.dataclass.BookclubMember
 import com.example.fe.R
 import com.example.fe.bookclub_book.adapter.BookclubByMonthRVAdapter
 import com.example.fe.bookclub_book.adapter.BookclubMemberRVAdapter
 import com.example.fe.databinding.FragmentBookclubBookHomeBinding
-import okhttp3.Response
 
 
 class BookclubBookHomeFragment : Fragment() {
@@ -41,12 +39,8 @@ class BookclubBookHomeFragment : Fragment() {
 
         val bookclubMemberRVAdapter = BookclubMemberRVAdapter(object : BookclubMemberRVAdapter.MyItemClickListener {
             override fun onItemClick(participation: ArrayList<BookclubMember>) {
-                // Fragment 전환
-                val detailFragment = BookclubBookReviewFragment()
-                parentFragmentManager.commit {
-                    replace(R.id.fragment_container, detailFragment)
-                    addToBackStack(null)
-                }
+                val intent = Intent(context, BookclubBookReview::class.java)
+                startActivity(intent)
             }
         })
 
@@ -69,8 +63,7 @@ class BookclubBookHomeFragment : Fragment() {
 
         val bookclubByMonthRVAdapter = BookclubByMonthRVAdapter(object : BookclubByMonthRVAdapter.MyItemClickListener {
             override fun onItemClick(participation: ArrayList<BookclubByMonth>) {
-                // BookClubJoinActivity로 전환
-                val intent = Intent(context, BookClubJoinActivity::class.java)
+                val intent = Intent(context, BookclubJoin::class.java)
                 startActivity(intent)
             }
         })
