+++ app/src/main/java/com/example/fe/mypage/MyPageFragment.kt
@@ -7,9 +7,9 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.widget.ViewPager2
-import com.example.fe.setting.SettingActivity
 import com.example.fe.databinding.FragmentMypageBinding
 import com.example.fe.mypage.adapter.MyPageVPAdapter
+import com.example.fe.Setting
 import com.google.android.material.tabs.TabLayout
 import com.google.android.material.tabs.TabLayoutMediator
 
@@ -43,7 +43,7 @@ class MyPageFragment : Fragment() {
         // Settings button click listener
         binding.mypageSettingIv.setOnClickListener {
             context?.let { ctx ->
-                val intent = Intent(ctx, SettingActivity::class.java)
+                val intent = Intent(ctx, Setting::class.java)
                 startActivity(intent)
             }
         }
