+++ app/src/main/res/layout/fragment_bookclub_book.xml
@@ -2,17 +2,72 @@
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@color/black"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:background="@color/colorMainBg"
+    android:id="@+id/fragment_container">
 
     <TextView
+        android:id="@+id/bookclub_book_title_tv"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="bookclub_book"
-        android:textColor="@color/white"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
+        android:text="책마다 북클럽"
+        android:textSize="20sp"
+        android:textStyle="bold"
+        android:textColor="@color/colorLabelNormal"
+        android:layout_marginTop="52dp"
+        android:layout_marginStart="20dp"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent" />
+        app:layout_constraintStart_toStartOf="parent" />
+
+    <ImageView
+        android:id="@+id/bookclub_book_search_ic"
+        android:layout_width="28dp"
+        android:layout_height="28dp"
+        android:src="@drawable/ic_magnifying_glass"
+        android:layout_marginEnd="16dp"
+        app:layout_constraintEnd_toStartOf="@+id/bookclub_book_person_ic"
+        app:layout_constraintTop_toTopOf="@+id/bookclub_book_title_tv"
+        app:layout_constraintBottom_toBottomOf="@+id/bookclub_book_title_tv" />
+
+    <ImageView
+        android:id="@+id/bookclub_book_person_ic"
+        android:layout_width="28dp"
+        android:layout_height="28dp"
+        android:src="@drawable/img_person_icon"
+        android:background="@drawable/circle_background"
+        android:scaleType="centerCrop"
+        android:clipToOutline="true"
+        android:layout_marginStart="16dp"
+        android:layout_marginEnd="20dp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="@+id/bookclub_book_search_ic"
+        app:layout_constraintBottom_toBottomOf="@+id/bookclub_book_search_ic" />
+
+    <com.google.android.material.tabs.TabLayout
+        android:id="@+id/bookclub_book_content_tl"
+        android:background="@color/colorMainBg"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="16dp"
+        app:tabTextColor="@color/colorLabelAlternative"
+        app:tabSelectedTextColor="@color/colorLabelNormal"
+        app:tabIndicatorColor="@color/colorLabelNormal"
+        app:tabIndicatorHeight="2dp"
+        app:tabIndicatorFullWidth="false"
+        app:tabGravity="fill"
+        app:tabMode="scrollable"
+        app:layout_constraintTop_toBottomOf="@id/bookclub_book_title_tv"
+        app:layout_constraintStart_toStartOf="parent" />
+
+    <androidx.viewpager2.widget.ViewPager2
+        android:id="@+id/bookclub_book_content_vp"
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        android:layout_marginTop="10dp"
+        app:layout_constraintVertical_bias="1.0"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/bookclub_book_content_tl" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
